requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.5
pandas>=1.3.5
matplotlib>=3.5.1
scikit-learn>=1.0.2
streamlit>=1.10.0
pillow>=9.0.0
opencv-python>=4.5.5 